# Use Rails default options for Haml's default options