## Changes
- To reduce object allocations on runtime, I added `Haml::Options.buffer_defaults` to use in `Haml::Buffer#initialize`.
- To avoid allocating new Hash, I changed `Haml::Buffer#initialize` not to call `Hash#merge` if no options are given.
- ~~Use `escape_html: true, ugly: true` as Haml's default options. This change gives significant performance gain for all Rails applications.~~
  - ~~related to: https://github.com/haml/haml/issues/894~~
  - ~~**This is breaking change.** I think we should ship this change in Haml 6.~~

## Details
Preamble code will be changed like:

```diff
-begin;extend Haml::Helpers;_hamlout = @haml_buffer = Haml::Buffer.new(haml_buffer, {:autoclose=>["area", "base", "basefont", "br", "col", "command", "embed", "frame", "hr", "img", "input", "isindex", "keygen", "link", "menuitem", "meta", "param", "source", "track", "wbr"], :preserve=>["textarea", "pre", "code"], :attr_wrapper=>"'", :ugly=>true, :format=>:html5, :encoding=>"UTF-8", :escape_html=>true, :escape_attrs=>true, :hyphenate_data_attrs=>true, :cdata=>false});_erbout = _hamlout.buffer;;_hamlout.buffer << ("<!DOCTYPE html>\n<html>\n<head>\n<title>Simple Benchmark</title>\n</head>\n<body>\n<h1>".freeze);; _hamlout.buffer << (
+begin;extend Haml::Helpers;_hamlout = @haml_buffer = Haml::Buffer.new(haml_buffer, {});_erbout = _hamlout.buffer;;_hamlout.buffer << ("<!DOCTYPE html>\n<html>\n<head>\n<title>Simple Benchmark</title>\n</head>\n<body>\n<h1>".freeze);; _hamlout.buffer << (

```


## Benchmark
With Ruby 2.4.0 and [k0kubun/haml_bench/templates/view.haml](https://github.com/k0kubun/haml_bench/blob/4a435313c94fe8f5131a870db2d145a290fdccb3/templates/view.haml),
### before
```
$ bundle exec ruby bench.rb
Calculating -------------------------------------
          haml 4.0.7     3.606k i/100ms
   haml 5.0.0.beta.2     3.969k i/100ms
-------------------------------------------------
          haml 4.0.7     36.977k (± 2.7%) i/s -    187.512k
   haml 5.0.0.beta.2     40.408k (± 2.2%) i/s -    202.419k

Comparison:
   haml 5.0.0.beta.2:    40408.4 i/s
          haml 4.0.7:    36977.4 i/s - 1.09x slower
```

### after
```
$ bundle exec ruby bench.rb
Calculating -------------------------------------
          haml 4.0.7     3.602k i/100ms
   haml 5.0.0.beta.2     4.764k i/100ms
-------------------------------------------------
          haml 4.0.7     36.435k (± 2.7%) i/s -    183.702k
   haml 5.0.0.beta.2     49.756k (± 2.5%) i/s -    252.492k

Comparison:
   haml 5.0.0.beta.2:    49756.4 i/s
          haml 4.0.7:    36435.3 i/s - 1.37x slower
```