@k0kubun 2017-02-06
I'll fix tests and merge this after Haml 5 is released.