@amatsuda 2017-02-07
How about updating the Rails specific default values in haml/railtie.rb?
We can keep the current behavior for both Rails users and non-Rails users that way (I'm not sure how hard it is for us to maintain both modes though).