@k0kubun 2017-02-07
> How about updating the Rails specific default values in haml/railtie.rb?

What does "updating the Rails specific default values" mean? I mean, what kind of code you intend to generate in that case?

If we don't alter global `Haml::Options.defaults` and have different `Haml::Template.options` or some Rails specific defeault vaules for Rails, Haml in this branch will generate `Haml::Buffer.new(haml_buffer, { escape_html: true, ugly: true })`. If the options are not empty, performance won't be optimized.

If we really want to keep double default values without performance regression, we should have `Haml::RailsBuffer` which has different buffer options.