@amatsuda 2017-02-07
> What does "updating the Rails specific default values" mean? I mean, what kind of code you intend to generate in that case?

This is what I meant.
https://github.com/amatsuda/haml/commit/122433db7495200a07d5518586f6f2ba5924fc59