@k0kubun 2017-02-07
With this patch, I confirmed `Haml::Buffer.new(haml_buffer, {})` is generated on `RAILS_ENV=production rails s`. This PR optimizes all Rails applications and changes no behavior. Merging.